requirements.txt
absl-py==0.8.0
alphapose==0.3.0+ddaf4b9
astor==0.8.0
certifi==2021.5.30
cloudpickle==2.1.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
freetype-py==2.2.0
google-pasta==0.1.7
gym==0.21.0
h5py==2.10.0
imageio==2.14.1
importlib-metadata==4.8.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
Markdown==3.1.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
olefile==0.46
Pillow==8.3.2
pip==21.3.1
pyglet==1.5.21
PyGLM==1.99.3
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.39
python-dateutil==2.8.2
scipy==1.5.4
setuptools==49.6.0.post20210108
six==1.12.0
termcolor==1.1.0
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
trimesh==3.8.18
typing_extensions==4.0.1
Werkzeug==0.16.0
wheel==0.33.6
wincertstore==0.2
wrapt==1.11.2
zipp==3.6.0
